requirement.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-oauth-toolkit==1.6.3
djangorestframework==3.15.2
idna==3.7
jwcrypto==1.5.6
oauth2-provider==0.0
oauthlib==3.2.2
pycparser==2.22
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
